Allow to choose image format for captures
