+++ src/frac_logic/mod.rs
@@ -1,6 +1,8 @@
 mod canvas_coords;
 mod fractal_logic;
-mod gpu;
+pub(crate) mod gpu_render;
+pub(crate) mod gpu_util;
+pub(crate) mod gpu_init;
 mod render_settings;
 
 pub(crate) use canvas_coords::CanvasCoords;
